Fixed 'stop' issue. If a function returns a dataset (like create_dist_matrix) it will cause the app to crash if it does not return "NULL" when it stops. This behavior had to be passed down to make_model_design and discretefish_subroutine
find_centroid was turning lat/lon point to characters. Had to convert back to numeric

map_viewer will only work if the map_viewer files are in the fishset package folder.  I revised the code to save files to the mapviewer folder in the fishst package folder but to also save files in the fishsetFolder